# backend/docs/swagger.yaml
basePath: /
definitions:
  activity.Activity:
    properties:
      activity_image:
        type: string
      creator_id:
        type: string
      date:
        type: string
      description:
        type: string
      id:
        type: string
      title:
        type: string
    type: object
  comment.Comment:
    properties:
      activity_id:
        type: string
      content:
        type: string
      created_at:
        type: string
      id:
        type: string
      user_id:
        type: string
    type: object
...
  responses.ActivityCreateRequest:
    properties:
      activity_image:
        example: https://example.com/image.jpg
        type: string
      date:
        example: "2025-12-31"
        type: string
      description:
        example: A competitive programming contest
        type: string
      title:
        example: Algorithm Contest
        type: string
    type: object
  responses.ActivityDeleteRequest:
    properties:
      creator_id:
        example: user123
        type: string
    type: object
  responses.ActivityUpdateRequest:
    properties:
      activity_image:
        example: https://example.com/image.jpg
        type: string
      date:
        example: "2025-12-31"
        type: string
      description:
        example: Updated description
        type: string
    type: object
...
  responses.CommentCreateRequest:
    properties:
      content:
        example: Great activity!
        type: string
      user_id:
        example: user123
        type: string
    type: object
  responses.CommentDeleteRequest:
    properties:
      requester_id:
        example: user123
        type: string
    type: object
  responses.CommentDeleteResponse:
    properties:
      comment_id:
        example: comment123
        type: string
      message:
        example: Comment deleted successfully
        type: string
    type: object
  responses.CommentsResponse:
    properties:
      activity_id:
        example: activity123
        type: string
      comment_count:
        example: 5
        type: integer
      comments:
        items:
          $ref: '#/definitions/comment.Comment'
        type: array
    type: object
...
  responses.ErrorResponse:
    properties:
      error:
        example: Invalid request
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: API for CODECK - COmpetição DE Código do Kim. A platform for competitive
    programmers.
  license:
    name: MIT
    url: https://github.com/your-username/codeck/blob/main/LICENSE
  termsOfService: http://swagger.io/terms/
  title: CODECK API
  version: "1.0"
paths:
  /activities:
    post:
      consumes:
      - application/json
      description: Create a new activity with title, date, and optional image/description
      parameters:
      - description: Activity creation data
        in: body
        name: activity
        required: true
        schema:
          $ref: '#/definitions/responses.ActivityCreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/activity.Activity'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
      summary: Create a new activity
      tags:
      - activities
  /activities/{activity_id}/comments:
    get:
      consumes:
      - application/json
      description: Get all comments for a specific activity
      parameters:
      - description: Activity ID
        in: path
        name: activity_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.CommentsResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
      summary: Get comments by activity ID
      tags:
      - comments
    post:
      consumes:
      - application/json
      description: Create a new comment on an activity
      parameters:
      - description: Activity ID
        in: path
        name: activity_id
        required: true
        type: string
      - description: Comment creation data
        in: body
        name: comment
        required: true
        schema:
          $ref: '#/definitions/responses.CommentCreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/comment.Comment'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
      summary: Create a new comment
      tags:
      - comments
  /activities/{id}:
    delete:
      consumes:
      - application/json
      description: Delete an activity (only creator can delete)
      parameters:
      - description: Activity ID
        in: path
        name: id
        required: true
        type: string
      - description: Delete request with creator_id
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/responses.ActivityDeleteRequest'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
      summary: Delete an activity
      tags:
      - activities
    get:
      consumes:
      - application/json
      description: Get activity information by activity ID
      parameters:
      - description: Activity ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/activity.Activity'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
      summary: Get activity by ID
      tags:
      - activities
    put:
      consumes:
      - application/json
      description: Update activity information (title cannot be updated)
      parameters:
      - description: Activity ID
        in: path
        name: id
        required: true
        type: string
      - description: Activity update data
        in: body
        name: activity
        required: true
        schema:
          $ref: '#/definitions/responses.ActivityUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/activity.Activity'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
      summary: Update an existing activity
      tags:
      - activities
  /comments/{comment_id}:
    delete:
      consumes:
      - application/json
      description: Delete a comment (only comment author or activity creator can delete)
      parameters:
      - description: Comment ID
        in: path
        name: comment_id
        required: true
        type: string
      - description: Delete request with requester_id
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/responses.CommentDeleteRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.CommentDeleteResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
      summary: Delete a comment
      tags:
      - comments
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"